requirements.txt
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
cryptography==37.0.2
Deprecated==1.2.13
dpcpp-cpp-rt==2022.1.0
filelock==3.6.0
Flask==2.1.2
Flask-Limiter==2.4.5.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
limits==2.6.2
MarkupSafe==2.1.0
mkl==2022.1.0
mkl-fft==1.3.1
mkl-service==2.4.0
msal==1.17.0
msal-extensions==1.0.0
numpy==1.22.3
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
opencensus-ext-flask==0.8.0
packaging==21.3
pandas==1.4.2
portalocker==2.4.0
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
rich==12.4.1
rsa==4.8
six==1.16.0
tbb==2021.6.0
typing_extensions==4.2.0
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.8.0
